% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_time_simulation.R
\name{one_time_simulation}
\alias{one_time_simulation}
\title{One time simulation}
\usage{
one_time_simulation(
  n_sim,
  beta_0,
  beta_1,
  beta_2,
  mu,
  sigma,
  na,
  nb,
  permuted = FALSE
)
}
\arguments{
\item{n_sim}{number of simulations}

\item{beta_0}{coefficient of the treatment}

\item{beta_1}{coefficient of the covariate 1}

\item{beta_2}{cofficient of the covariate 2}

\item{mu}{the overall mean response}

\item{sigma}{standard error of the error term}

\item{na}{number of patients in the control}

\item{nb}{number of patients in the treatment}

\item{permuted}{whether use the permutation algorithm (Default is FALSE)}
}
\value{
data_sim
}
\description{
One time simulation
}
