% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_zhang.R
\name{simulation_zhang}
\alias{simulation_zhang}
\title{Simulation for the prespecified iterations}
\usage{
simulation_zhang(
  n_sim,
  parallel_state = FALSE,
  random_seed = 123,
  beta_0,
  beta_1,
  beta_2,
  mu,
  sigma,
  na,
  nb,
  permuted = FALSE
)
}
\arguments{
\item{n_sim}{number of simulations}

\item{parallel_state}{whether to use the parallel function (Default is False)}

\item{random_seed}{specify a random seed to reproduce the results (Default seed is 123)}

\item{beta_0}{coefficient of the treatment}

\item{beta_1}{coefficient of the covariate 1}

\item{beta_2}{cofficient of the covariate 2}

\item{mu}{the overall mean response}

\item{sigma}{standard error of the error term}

\item{na}{number of patients in the control}

\item{nb}{number of patients in the treatment}

\item{permuted}{whether use the permutation algorithm (Default is FALSE)}
}
\value{
data_sim
}
\description{
Simulation for the prespecified iterations
}
