% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_corr.R
\name{summarize_corr}
\alias{summarize_corr}
\title{Calculation the correlation between the prespecified test statistics}
\usage{
summarize_corr(data_sim)
}
\arguments{
\item{data_sim}{the simulated data}
}
\value{
Pearson correlation coefficient between test statistics
}
\description{
Calculation the correlation between the prespecified test statistics
}
