% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_power.R
\name{summarize_power}
\alias{summarize_power}
\title{Calculation of Power for tests}
\usage{
summarize_power(data_sim, crit_val, signi_level = 0.05, permuted = FALSE)
}
\arguments{
\item{data_sim}{the simulated data}

\item{crit_val}{identified critical value of minP from 'summarize_type1' function}

\item{signi_level}{significance level (default alpha=0.05)}

\item{permuted}{whether the simulated data includes the minP via permutation (default is FALSE)}
}
\value{
value of power for minP test, and prespecified model 1 and model 2.
}
\description{
Calculation of Power for tests
}
