% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_proportion.R
\name{summarize_proportion}
\alias{summarize_proportion}
\title{Calculation the proportion of minP choosing model 1}
\usage{
summarize_proportion(data_sim)
}
\arguments{
\item{data_sim}{the simulated data}
}
\value{
Proportion that the minP test statistic chooses
}
\description{
Calculation the proportion of minP choosing model 1
}
