% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_type1.R
\name{summarize_type1}
\alias{summarize_type1}
\title{Calculation of Type 1 error and critical values of tests}
\usage{
summarize_type1(data_sim, signi_level = 0.05, permuted = FALSE)
}
\arguments{
\item{data_sim}{the simulated data}

\item{signi_level}{significance level (default alpha=0.05)}

\item{permuted}{whether the simulated data includes the minP via permutation (default is FALSE)}
}
\value{
list of type 1 errors and critical values
}
\description{
Calculation of Type 1 error and critical values of tests
}
